Novos testes

+++ src/Fieldclimate.Pessl.Domain/Model/Vals.cs
@@ -2,7 +2,7 @@ namespace Fieldclimate.Pessl.Domain.Model
 {
     public class Vals
     {
-        public int min { get; set; }
-        public int max { get; set; }
+        public long min { get; set; }
+        public long max { get; set; }
     }
 }
